move demo middleware
